Structure Refactoring: Phase 02
