+++ lib/domain/models/media_entity_collection.dart
@@ -9,11 +9,11 @@ import '../../shared/concurrency_manager.dart';
 import '../entities/media_entity.dart';
 import '../services/core/logging_service.dart';
 import '../services/core/service_container.dart';
-import '../services/metadata/coordinate_extraction/exif_coordinate_extractor.dart';
-import '../services/metadata/date_extraction/exif_date_extractor.dart';
-import '../services/metadata/date_extraction/json_date_extractor.dart';
-import '../services/metadata/exif_writer_service.dart';
-import '../services/metadata/json_metadata_matcher_service.dart';
+import '../../steps/step_04_extract_dates/services/gps_extractors/exif_gps_extractor.dart';
+import '../../steps/step_04_extract_dates/services/date_extractors/exif_date_extractor.dart';
+import '../../steps/step_04_extract_dates/services/date_extractors/json_date_extractor.dart';
+import '../../steps/step_05_write_exif/services/exif_writer_service.dart';
+import '../../steps/step_04_extract_dates/services/json_metadata_matcher_service.dart';
 import '../value_objects/date_time_extraction_method.dart';
 
 /// Modern domain model representing a collection of media entities.
