+++ lib/shared/exports.dart
@@ -8,7 +8,7 @@ library;
 export '../domain/services/core/logging_service.dart';
 export '../domain/services/core/service_container.dart';
 export '../domain/services/media/mime_type_service.dart';
-export '../domain/services/metadata/json_metadata_matcher_service.dart';
+export '../steps/step_04_extract_dates/services/json_metadata_matcher_service.dart';
 // Concurrency management
 export 'concurrency_manager.dart';
 // Application constants
