+++ lib/steps/step_02_discover_media/step_02_discover_media.dart
@@ -7,7 +7,7 @@ import '../../shared/extensions/file_extensions.dart';
 import '../../domain/entities/media_entity.dart';
 import '../../domain/models/pipeline_step_model.dart';
 import '../../domain/models/processing_config_model.dart';
-import '../../domain/services/metadata/date_extraction/json_date_extractor.dart';
+import '../step_04_extract_dates/services/date_extractors/json_date_extractor.dart';
 import '../../domain/services/processing/edited_version_detector_service.dart';
 import '../../domain/services/processing/takeout_folder_classifier_service.dart';
