+++ lib/steps/step_04_extract_dates/services/date_extractors/exif_date_extractor.dart
@@ -10,9 +10,9 @@ import 'package:mime/mime.dart';
 import '../../../../infrastructure/exiftool_service.dart';
 import '../../../../shared/constants.dart';
 import '../../../../shared/constants/exif_constants.dart';
-import '../../core/global_config_service.dart';
-import '../../core/logging_service.dart';
-import '../../core/service_container.dart'; // <-- Added to check global dictionary presence
+import '../../../../domain/services/core/global_config_service.dart';
+import '../../../../domain/services/core/logging_service.dart';
+import '../../../../domain/services/core/service_container.dart'; // <-- Added to check global dictionary presence
 
 class _SmartReadResult {
   _SmartReadResult(this.bytes, this.usedHeadOnly);
