+++ lib/steps/step_07_move_files/services/file_operation_service.dart
@@ -3,10 +3,10 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
-import '../../../../shared/concurrency_manager.dart';
-import '../../../../shared/global_pools.dart';
-import '../../core/logging_service.dart';
-import '../../core/service_container.dart';
+import '../../../shared/concurrency_manager.dart';
+import '../../../shared/global_pools.dart';
+import '../../../domain/services/core/logging_service.dart';
+import '../../../domain/services/core/service_container.dart';
 
 /// High-performance file operation service with optimized I/O and concurrency control
 ///
