+++ lib/steps/step_07_move_files/services/path_generator_service.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:path/path.dart' as p;
 
-import '../../../models/processing_config_model.dart';
+import '../../../domain/models/processing_config_model.dart';
 import 'moving_context_model.dart';
 
 /// Service responsible for generating file and directory paths
