+++ lib/steps/step_07_move_files/step_07_move_files.dart
@@ -3,9 +3,9 @@ import 'package:console_bars/console_bars.dart';
 
 import '../../domain/entities/media_entity.dart';
 import '../../domain/models/pipeline_step_model.dart';
-import '../../domain/services/file_operations/moving/media_entity_moving_service.dart';
-import '../../domain/services/file_operations/moving/moving_context_model.dart';
-import '../../domain/services/file_operations/moving/strategies/media_entity_moving_strategy.dart';
+import 'services/media_entity_moving_service.dart';
+import 'services/moving_context_model.dart';
+import 'strategies/media_entity_moving_strategy.dart';
 import '../../domain/value_objects/media_files_collection.dart';
 
 /// Step 7: Move files to output directory
