+++ lib/steps/step_07_move_files/strategies/media_entity_moving_strategy_factory.dart
@@ -1,7 +1,7 @@
-import '../../../../models/processing_config_model.dart';
-import '../file_operation_service.dart';
-import '../path_generator_service.dart';
-import '../symlink_service.dart';
+import '../../../domain/models/processing_config_model.dart';
+import '../services/file_operation_service.dart';
+import '../services/path_generator_service.dart';
+import '../services/symlink_service.dart';
 import 'duplicate_copy_moving_strategy.dart';
 import 'json_moving_strategy.dart';
 import 'media_entity_moving_strategy.dart';
