+++ lib/steps/step_07_move_files/strategies/shortcut_moving_strategy.dart
@@ -1,106 +1,73 @@
 import 'dart:io';
 
-import '../../../../entities/media_entity.dart';
-import '../file_operation_service.dart';
-import '../moving_context_model.dart';
-import '../path_generator_service.dart';
+import '../../../domain/entities/media_entity.dart';
+import '../services/file_operation_service.dart';
+import '../services/moving_context_model.dart';
+import '../services/path_generator_service.dart';
+import '../services/symlink_service.dart';
 import 'media_entity_moving_strategy.dart';
 
-/// Duplicate-copy moving strategy implementation
+/// Shortcut moving strategy implementation
 ///
-/// This strategy creates actual file copies in both ALL_PHOTOS and album folders.
-/// Files are moved to ALL_PHOTOS and copied to each album folder.
-class DuplicateCopyMovingStrategy extends MediaEntityMovingStrategy {
-  const DuplicateCopyMovingStrategy(this._fileService, this._pathService);
+/// This strategy creates shortcuts/symlinks from album folders to files in ALL_PHOTOS.
+/// Files are moved to ALL_PHOTOS organized by date, and shortcuts are created in album folders.
+class ShortcutMovingStrategy extends MediaEntityMovingStrategy {
+  const ShortcutMovingStrategy(
+    this._fileService,
+    this._pathService,
+    this._symlinkService,
+  );
 
   final FileOperationService _fileService;
   final PathGeneratorService _pathService;
+  final SymlinkService _symlinkService;
 
   @override
-  String get name => 'Duplicate Copy';
+  String get name => 'Shortcut';
 
   @override
-  bool get createsShortcuts => false;
+  bool get createsShortcuts => true;
 
   @override
-  bool get createsDuplicates => true;
+  bool get createsDuplicates => false;
 
   @override
   Stream<MediaEntityMovingResult> processMediaEntity(
     final MediaEntity entity,
     final MovingContext context,
   ) async* {
-    // Step 1: Ensure we have (or create) the primary canonical file in ALL_PHOTOS (or PARTNER_SHARED)
-    final originalPrimaryFile = entity.primaryFile;
+    final results = <MediaEntityMovingResult>[];
+
+    // Step 1: Move primary file to ALL_PHOTOS (or PARTNER_SHARED)
+    final primaryFile = entity.primaryFile;
     final allPhotosDir = _pathService.generateTargetDirectory(
-      null, // null = ALL_PHOTOS or PARTNER_SHARED
+      null, // null = ALL_PHOTOS (or PARTNER_SHARED if partner shared)
       entity.dateTaken,
       context,
       isPartnerShared: entity.partnershared,
     );
-
-    // If the original file physically resides inside one of the album folders that we will later populate,
-    // we must MOVE it to the canonical ALL_PHOTOS folder instead of copying it, otherwise the original
-    // remains and causes leftover files. Heuristic: if its path contains any of the album directory target paths.
-    File canonicalFile = originalPrimaryFile;
-    // Track canonical file; no need for additional collections here.
     final stopwatch = Stopwatch()..start();
     try {
-      if (entity.albumNames.isNotEmpty) {
-        // Pre-generate album directories to test path containment cheaply
-        final albumDirs = <String, Directory>{};
-        for (final albumName in entity.albumNames) {
-          final dir = _pathService.generateTargetDirectory(
-            albumName,
-            entity.dateTaken,
-            context,
-            isPartnerShared: entity.partnershared,
-          );
-          albumDirs[albumName] = dir;
-        }
-
-        final originalPathLower = originalPrimaryFile.path.toLowerCase();
-        final isInsideAlbum = albumDirs.values.any(
-          (final d) => originalPathLower.startsWith(d.path.toLowerCase()),
-        );
-
-        if (isInsideAlbum) {
-          // Move the file out to canonical location
-          canonicalFile = await _fileService.moveFile(
-            originalPrimaryFile,
-            allPhotosDir,
-            dateTaken: entity.dateTaken,
-          );
-        } else {
-          // Regular flow: move (rename) into ALL_PHOTOS (if already there rename handles uniqueness)
-          canonicalFile = await _fileService.moveFile(
-            originalPrimaryFile,
-            allPhotosDir,
-            dateTaken: entity.dateTaken,
-          );
-        }
-      } else {
-        // No albums: just move
-        canonicalFile = await _fileService.moveFile(
-          originalPrimaryFile,
-          allPhotosDir,
-          dateTaken: entity.dateTaken,
-        );
-      }
-
+      final movedFile = await _fileService.moveFile(
+        primaryFile,
+        allPhotosDir,
+        dateTaken: entity.dateTaken,
+      );
       stopwatch.stop();
-      yield MediaEntityMovingResult.success(
+      final primaryResult = MediaEntityMovingResult.success(
         operation: MediaEntityMovingOperation(
-          sourceFile: originalPrimaryFile,
+          sourceFile: primaryFile,
           targetDirectory: allPhotosDir,
           operationType: MediaEntityOperationType.move,
           mediaEntity: entity,
         ),
-        resultFile: canonicalFile,
+        resultFile: movedFile,
         duration: stopwatch.elapsed,
       );
+      results.add(primaryResult);
+      yield primaryResult;
 
-      // Step 2: For each album, ensure a copy exists (unless the original file was already there and moved)
+      // Step 2: Create shortcuts for each album association
       for (final albumName in entity.albumNames) {
         final albumDir = _pathService.generateTargetDirectory(
           albumName,
@@ -108,53 +75,58 @@ class DuplicateCopyMovingStrategy extends MediaEntityMovingStrategy {
           context,
           isPartnerShared: entity.partnershared,
         );
-
-        final copyStopwatch = Stopwatch()..start();
+        final shortcutStopwatch = Stopwatch()..start();
         try {
-          final copiedFile = await _fileService.copyFile(
-            canonicalFile,
+          final shortcutFile = await _symlinkService.createSymlink(
             albumDir,
-            dateTaken: entity.dateTaken,
+            movedFile,
           );
-          copyStopwatch.stop();
-          yield MediaEntityMovingResult.success(
+
+          shortcutStopwatch.stop();
+          final shortcutResult = MediaEntityMovingResult.success(
             operation: MediaEntityMovingOperation(
-              sourceFile: canonicalFile,
+              sourceFile: movedFile,
               targetDirectory: albumDir,
-              operationType: MediaEntityOperationType.copy,
+              operationType: MediaEntityOperationType.createSymlink,
               mediaEntity: entity,
               albumKey: albumName,
             ),
-            resultFile: copiedFile,
-            duration: copyStopwatch.elapsed,
+            resultFile: shortcutFile,
+            duration: shortcutStopwatch.elapsed,
           );
+          results.add(shortcutResult);
+          yield shortcutResult;
         } catch (e) {
-          copyStopwatch.stop();
-          yield MediaEntityMovingResult.failure(
+          shortcutStopwatch.stop();
+          final errorResult = MediaEntityMovingResult.failure(
             operation: MediaEntityMovingOperation(
-              sourceFile: canonicalFile,
+              sourceFile: movedFile,
               targetDirectory: albumDir,
-              operationType: MediaEntityOperationType.copy,
+              operationType: MediaEntityOperationType.createSymlink,
               mediaEntity: entity,
               albumKey: albumName,
             ),
-            errorMessage: 'Failed to copy to album folder: $e',
-            duration: copyStopwatch.elapsed,
+            errorMessage: 'Failed to create shortcut: $e',
+            duration: shortcutStopwatch.elapsed,
           );
+          results.add(errorResult);
+          yield errorResult;
         }
       }
     } catch (e) {
       stopwatch.stop();
-      yield MediaEntityMovingResult.failure(
+      final errorResult = MediaEntityMovingResult.failure(
         operation: MediaEntityMovingOperation(
-          sourceFile: originalPrimaryFile,
+          sourceFile: primaryFile,
           targetDirectory: allPhotosDir,
           operationType: MediaEntityOperationType.move,
           mediaEntity: entity,
         ),
-        errorMessage: 'Failed to establish canonical file: $e',
+        errorMessage: 'Failed to move primary file: $e',
         duration: stopwatch.elapsed,
       );
+      results.add(errorResult);
+      yield errorResult;
     }
 
     // NEW: move non-primary physical files to _Duplicates preserving source structure
@@ -163,7 +135,7 @@ class DuplicateCopyMovingStrategy extends MediaEntityMovingStrategy {
 
   @override
   void validateContext(final MovingContext context) {
-    // No special validation needed for duplicate-copy strategy
+    // No special validation needed for shortcut mode
   }
 
   // --- NEW helper: move all non-primary physical files to _Duplicates, preserving structure ---
@@ -226,7 +198,6 @@ class DuplicateCopyMovingStrategy extends MediaEntityMovingStrategy {
     }
   }
 
-  // Calculate a relative directory inside _Duplicates that mirrors the source tree.
   _RelInfo _computeDuplicatesRelativeInfo(final String sourcePath) {
     final normalized = sourcePath.replaceAll('\\', '/');
     final lower = normalized.toLowerCase();
