+++ test/integration/moving_integration_test.dart
@@ -7,8 +7,8 @@ import 'package:gpth/domain/entities/media_entity.dart';
 import 'package:gpth/domain/models/media_entity_collection.dart';
 import 'package:gpth/domain/models/processing_config_model.dart';
 import 'package:gpth/domain/services/core/service_container.dart';
-import 'package:gpth/domain/services/file_operations/moving/media_entity_moving_service.dart';
-import 'package:gpth/domain/services/file_operations/moving/moving_context_model.dart';
+import 'package:gpth/steps/step_07_move_files/services/media_entity_moving_service.dart';
+import 'package:gpth/steps/step_07_move_files/services/moving_context_model.dart';
 import 'package:test/test.dart';
 
 import '../setup/test_setup.dart';
