+++ test/unit/date_extractors_test.dart
@@ -10,7 +10,7 @@ import 'dart:io';
 import 'package:coordinate_converter/coordinate_converter.dart';
 import 'package:gpth/domain/services/core/global_config_service.dart';
 import 'package:gpth/domain/services/core/service_container.dart';
-import 'package:gpth/domain/services/metadata/date_extraction/date_extractor_service.dart';
+import 'package:gpth/steps/step_04_extract_dates/services/date_extractors/date_extractor_service.dart';
 import 'package:gpth/infrastructure/exiftool_service.dart';
 import 'package:test/test.dart';
