+++ test/unit/extension_fixing_metadata_matcher_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:gpth/domain/services/metadata/date_extraction/json_date_extractor.dart';
-import 'package:gpth/domain/services/metadata/json_metadata_matcher_service.dart';
+import 'package:gpth/steps/step_04_extract_dates/services/date_extractors/json_date_extractor.dart';
+import 'package:gpth/steps/step_04_extract_dates/services/json_metadata_matcher_service.dart';
 import 'package:test/test.dart';
 import '../setup/test_setup.dart';
