+++ test/unit/partner_sharing_integration_test.dart
@@ -2,8 +2,8 @@ import 'dart:io';
 
 import 'package:gpth/domain/entities/media_entity.dart';
 import 'package:gpth/domain/models/processing_config_model.dart';
-import 'package:gpth/domain/services/file_operations/moving/moving_context_model.dart';
-import 'package:gpth/domain/services/file_operations/moving/path_generator_service.dart';
+import 'package:gpth/steps/step_07_move_files/services/moving_context_model.dart';
+import 'package:gpth/steps/step_07_move_files/services/path_generator_service.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
