+++ test/unit/partner_sharing_test.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:gpth/domain/entities/media_entity.dart';
-import 'package:gpth/domain/services/metadata/date_extraction/json_date_extractor.dart';
+import 'package:gpth/steps/step_04_extract_dates/services/date_extractors/json_date_extractor.dart';
 import 'package:test/test.dart';
 
 import '../setup/test_setup.dart';
